Agregar y Reportar

+++ src/org/jorgeperalta/controller/MenuClientesController.java
@@ -18,6 +18,7 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
 import javafx.scene.control.TextField;
@@ -26,6 +27,7 @@ import org.jorgeperalta.dao.Conexion;
 import org.jorgeperalta.dto.ClienteDTO;
 import org.jorgeperalta.model.Cliente;
 import org.jorgeperalta.system.Main;
+import org.jorgeperalta.utils.SuperKinalAlert;
 
 /**
  * FXML Controller class
@@ -64,9 +66,11 @@ public class MenuClientesController implements Initializable {
         } else if (event.getSource() == btnRegresar){
             stage.menuPrincipalView();
         } else if(event.getSource() == btnEliminar){
-            int cliId = ((Cliente)tblClientes.getSelectionModel().getSelectedItem()).getClienteId();
-            eliminarCliente(cliId);
-            cargarLista();
+            if(SuperKinalAlert.getInstance().mostrarAlertaConfirmacion(405).get() == ButtonType.OK){
+                int cliId = ((Cliente)tblClientes.getSelectionModel().getSelectedItem()).getClienteId();
+                eliminarCliente(cliId);
+                cargarLista();
+            }
         } else if(event.getSource() == btnBuscar){
             tblClientes.getItems().clear();
             if(tfClienteId.getText().equals("")){
